Optimize Jaden Smith

+++ catscripts/Entertainment/Actors_and_actresses/Jaden_Smith/JadenSmith.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_JADENSMITH="Jaden(| )Smith"
+KEYWORDS_ACTOR_JADENSMITH="Jaden(| )Smith"
 
-egrep -i "$KEYWORDS_ACTOR_JADENSMITH" newpages.txt >> JadenSmith.txt
-
-JADENSMITH=`stat --print=%s JadenSmith.txt`
-
-if [ $JADENSMITH -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="JadenSmith.txt"
-  export CATNAME="Jaden Smith"
-  $CATEGORIZE
-fi
 
-rm JadenSmith.txt
+  JADENSMITH=`egrep -i "$KEYWORDS_ACTOR_JADENSMITH" newpages.txt`
+
+  if [ "$JADENSMITH" != "" ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_JADENSMITH" newpages.txt > JadenSmith.txt
+    export CATFILE="JadenSmith.txt"
+    export CATNAME="Jaden Smith"
+    $CATEGORIZE
+    rm JadenSmith.txt
+    unset JADENSMITH
+  fi
+fi
